# Add new `Performance/CollectionLiteralInMethod` cop