--- a/changelog/new_collection_literal_in_method.md
+++ b/changelog/new_collection_literal_in_method.md
@@ -1,0 +1,1 @@
+* [#303](https://github.com/rubocop/rubocop-performance/pull/303): Add new Performance/CollectionLiteralInMethod cop. ([@exoego][])