… lib/rubocop/cop/performance_cops.rb