@koic 2022-09-05
Hhm, I'm not sure if the benchmark you provided is reasonable. Because the benchmark does not include the cost of assignment to constants.
For methods that are executed only once, writing literals as they are seems faster than constant assignment. So, unlike `Performance/CollectionLiteralInLoop` cop, constants may not always be faster than literals.