@exoego 2022-09-05
> For methods that are executed only once, writing literals as they are seems faster than constant assignment. 

Correct.
I think #71 is for methods that are invoked many times.